+++ src/Compute/Compute.Test/ScenarioTests/AvailabilitySetTests.cs
@@ -51,5 +51,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-AvailabilitySetConvert 'eastus2euap'");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestAvailabilitySetScheduledEventsPolicy()
+        {
+            TestRunner.RunTestScript("Test-AvailabilitySetScheduledEventsPolicy 'eastus2euap'");
+        }
     }
 }