+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineTests.cs
@@ -703,5 +703,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-VirtualMachineGalleryApplicationFlags");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVirtualMachineScheduledEventsPolicy()
+        {
+            TestRunner.RunTestScript("Test-VirtualMachineScheduledEventsPolicy 'eastus2euap'");
+        }
     }
 }